Created file api_caller.py api_caller is a further abstraction on telegrambot that allows api_caller's calling file to get and send messages with one simple function call.
Created tester.py
A test bench to try out (call) the packages being created

Created message_processor.py
This file will process the user messages that will be sent to it,
and will decide a course of action for what to do next depending
on the message.
It may be updated later to also take the state of the current tran
-saction as an argument. For now. It is being designed to work with
one single user.

Modified telegrambot.py
---> Removed getMsg(), this functionality has been added ignored
api_caller.py
---> Removed some redundancies from receive().