Created a new file db_interface.py that handles the interaction of the app with the PostgreSQL database using psycopg2.
The file contains a class pool_init() and two standalone functions
cnnct() and dscnnct().

The pool_init class initializes a psycopg2 connection pool object
that can be used by the application to get connections from when
needed using the cnnct() function, and the connections can be put
back to standy by using the dscnnct() function. The dscnnct() func-
-tion also handles commits to the database as well as creation and
closing of cursors.

This addtion significantly increased the performance of the app in
terms of the quickness of the bot's replies.

The changes made to other parts of the application that also facil
-itate the above stated changes will be added in the next commit.